backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlmodel>=0.0.14
python-multipart>=0.0.6
python-docx>=1.1.0
pypdf>=3.17.0
chromadb>=0.4.22
python-dotenv>=1.0.1
aiofiles>=23.2.1
httpx>=0.27.0
psycopg2-binary>=2.9.9
pydantic>=2.0.0
pydantic-settings>=2.0.0
alembic>=1.13.0
reportlab>=4.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
